requirements.txt
aiogram==3.13.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.0
python-dotenv==1.0.1
